add api security

+++ src/main/java/io/vertx/guides/wiki/http/HttpServerVerticle.java
@@ -19,8 +19,13 @@ import io.vertx.core.http.HttpServerOptions;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import io.vertx.core.net.JksOptions;
+import io.vertx.ext.auth.KeyStoreOptions;
+import io.vertx.ext.auth.User;
 import io.vertx.ext.auth.jdbc.JDBCAuth;
+import io.vertx.ext.auth.jwt.JWTAuth;
+import io.vertx.ext.auth.jwt.JWTAuthOptions;
 import io.vertx.ext.jdbc.JDBCClient;
+import io.vertx.ext.jwt.JWTOptions;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
 import io.vertx.ext.web.client.HttpResponse;
@@ -31,6 +36,7 @@ import io.vertx.ext.web.handler.AuthHandler;
 import io.vertx.ext.web.handler.BodyHandler;
 import io.vertx.ext.web.handler.CookieHandler;
 import io.vertx.ext.web.handler.FormLoginHandler;
+import io.vertx.ext.web.handler.JWTAuthHandler;
 import io.vertx.ext.web.handler.RedirectAuthHandler;
 import io.vertx.ext.web.handler.SessionHandler;
 import io.vertx.ext.web.handler.UserSessionHandler;
@@ -107,6 +113,44 @@ public class HttpServerVerticle extends AbstractVerticle implements DatabaseCons
 		
 		
 		Router apiRouter = Router.router(vertx);
+		
+		
+		JWTAuth jwtAuth = JWTAuth.create(vertx, new JWTAuthOptions()
+				.setKeyStore(new KeyStoreOptions()
+						.setPath("keystore.jceks")
+						.setType("jceks")
+						.setPassword("secret")));
+		
+		apiRouter.route().handler(JWTAuthHandler.create(jwtAuth,"/api/token"));
+		
+		//generate token
+		apiRouter.get("/token").handler(context -> {
+			JsonObject creds = new JsonObject()
+					.put("username", context.request().getHeader("login"))
+					.put("password", context.request().getHeader("password"));
+			auth.authenticate(creds, authResult -> {
+				if(authResult.succeeded()) {
+					User user = authResult.result();
+					user.isAuthorised("create", canCreate -> {
+						user.isAuthorised("delete", canDelete -> {
+							user.isAuthorised("update", canUpdate -> {
+								String token = jwtAuth.generateToken(new JsonObject()
+										.put("username", context.request().getHeader("login"))
+										.put("canCreate", canCreate.succeeded() && canCreate.result())
+										.put("canDelete", canDelete.succeeded() && canDelete.result())
+										.put("canUpdate", canUpdate.succeeded() && canUpdate.result()), 
+										new JWTOptions().setSubject("Wiki API").setIssuer("Vert.x"));
+								context.response().putHeader("Content-type", "text/plain").end(token);
+							});
+						});
+					});
+				} else {
+					context.fail(401);
+				}
+			});
+		});
+		
+		
 		apiRouter.get("/pages").handler(this::apiRoot);
 		apiRouter.get("/pages/:id").handler(this::apiGetPage);
 		apiRouter.post().handler(BodyHandler.create());
@@ -407,23 +451,28 @@ public class HttpServerVerticle extends AbstractVerticle implements DatabaseCons
 	
 	
 	private void apiCreatePage(RoutingContext context) {
-		JsonObject page = context.getBodyAsJson();
-		if(!validateJsonPageDocument(context, page,"name","markdown")) {
-			return;
+		if(context.user().principal().getBoolean("canCreate",false)) {
+			JsonObject page = context.getBodyAsJson();
+			if(!validateJsonPageDocument(context, page,"name","markdown")) {
+				return;
+			}
+			
+			dbService.createPage(page.getString("name"), page.getString("markdown"), reply -> {
+				if(reply.succeeded()) {
+					context.response().setStatusCode(201);
+					context.response().putHeader("Content-Type", "application/json");
+					context.response().end(new JsonObject().put("success", true).encode());
+				} else {
+					context.response().setStatusCode(500);
+					context.response().putHeader("Content-Type", "application/json");
+					context.response().end(new JsonObject().put("success", false)
+							.put("error", reply.cause().getMessage()).encode());
+				}
+			});
+		}else {
+			context.fail(401);
 		}
 		
-		dbService.createPage(page.getString("name"), page.getString("markdown"), reply -> {
-			if(reply.succeeded()) {
-				context.response().setStatusCode(201);
-				context.response().putHeader("Content-Type", "application/json");
-				context.response().end(new JsonObject().put("success", true).encode());
-			} else {
-				context.response().setStatusCode(500);
-				context.response().putHeader("Content-Type", "application/json");
-				context.response().end(new JsonObject().put("success", false)
-						.put("error", reply.cause().getMessage()).encode());
-			}
-		});
 	}
 	
 	private boolean validateJsonPageDocument(RoutingContext context, JsonObject page, String...expectedKeys) {
@@ -441,14 +490,19 @@ public class HttpServerVerticle extends AbstractVerticle implements DatabaseCons
 	
 	
 	private void apiUpdatePage(RoutingContext context) {
-		int id = Integer.valueOf(context.request().getParam("id"));
-		JsonObject page = context.getBodyAsJson();
-		if(!validateJsonPageDocument(context, page, "markdown")) {
-			return;
+		if(context.user().principal().getBoolean("canUpdate",false)) {
+			int id = Integer.valueOf(context.request().getParam("id"));
+			JsonObject page = context.getBodyAsJson();
+			if(!validateJsonPageDocument(context, page, "markdown")) {
+				return;
+			}
+			dbService.savePage(id, page.getString("markdown"), reply -> {
+				handleSimpleDbReply(context, reply);
+			});
+		} else {
+			context.fail(401);
 		}
-		dbService.savePage(id, page.getString("markdown"), reply -> {
-			handleSimpleDbReply(context, reply);
-		});
+		
 	}
 	
 	private void handleSimpleDbReply(RoutingContext context, AsyncResult<Void> reply) {
@@ -466,10 +520,15 @@ public class HttpServerVerticle extends AbstractVerticle implements DatabaseCons
 	}
 	
 	private void apiDeletePage(RoutingContext context) {
-		int id = Integer.valueOf(context.request().getParam("id"));
-		dbService.deletePage(id, reply -> {
-			handleSimpleDbReply(context, reply);
-		});
+		if(context.user().principal().getBoolean("canDelete", false)) {
+			int id = Integer.valueOf(context.request().getParam("id"));
+			dbService.deletePage(id, reply -> {
+				handleSimpleDbReply(context, reply);
+			});
+		}else {
+			context.fail(401);
+		}
+		
 	}
 	
 	private void loginHandler(RoutingContext context) {
